fix: ModalTemplate and StartCurrentSession for Android.
