+++ packages/native-frontend/components/UtilComponents/ModalTemplate.tsx
@@ -1,0 +1,42 @@
+import {useState} from "react";
+import {
+    Modal,
+    Pressable,
+    View,
+    Text,
+    TouchableWithoutFeedback,
+} from "react-native";
+
+type Props = {
+    Icon: JSX.Element;
+    Content: (setModalVisible: (visible: boolean) => void) => JSX.Element;
+};
+
+export default function ModalTemplate({Icon, Content}: Props)
+{
+    const [modalVisible, setModalVisible] = useState<boolean>(false);
+
+    return (
+        <View>
+            <Pressable
+                onPress={() =>
+                {
+                    setModalVisible(true);
+                }}
+            >
+                {Icon}
+            </Pressable>
+            <Modal visible={modalVisible} transparent={true}>
+                <Pressable
+                    onPress={() => setModalVisible(false)}
+                    className="absolute top-0 left-0 right-0 bottom-0 bg-black/60"
+                />
+                <View className="flex-1 justify-center items-center">
+                    <View className="bg-white p-4 rounded-xl rounded-lg">
+                        {Content(setModalVisible)}
+                    </View>
+                </View>
+            </Modal>
+        </View>
+    );
+}
