+++ packages/native-frontend/components/currSession/StartCurrentSession.tsx
@@ -1,5 +1,5 @@
 import {FlatList, Pressable, View, Text, TextInput} from "react-native";
-import ModalTemplate from "../ModalTemplate";
+import ModalTemplate from "../UtilComponents/ModalTemplate";
 import {useTemplateContext} from "@/util/templateContext";
 import {useState} from "react";
 import {Template} from "@/types/template";
@@ -43,29 +43,33 @@ export default function StartCurrentSession({Icon}: Props)
             Content={(setModalVisible) => (
                 <View className="rounded-xl px-4 pt-4 pb-4 bg-white">
                     <Pressable onPress={startBlank}>
-                        <Text className="text-2xl px-3 py-1 rounded-xl outline outline-1 outline-gray-300 bg bg-gray-200 hover:bg-gray-300">
+                        <Text className="text-2xl px-3 py-1 rounded-xl outline outline-1 outline-gray-300 bg bg-gray-200">
                             Create Blank Session
                         </Text>
                     </Pressable>
                     <View className="py-1" />
-                    <View className="flex-row text-base px-3 py-1 rounded-xl outline outline-1 outline-gray-300 bg-gray-200">
-                        <Text>Templates: </Text>
+                    <View className="flex-row justify-start items-center px-3 py-1 rounded-xl outline outline-1 outline-gray-300 bg-gray-200">
+                        <Text className="text-lg">Templates: </Text>
                         <TextInput
+                            className="text-lg"
                             placeholder="Search"
                             value={search}
                             onChangeText={setSearch}
                             placeholderTextColor="#606060"
                         />
                     </View>
-                    <View className="max-h-28">
+                    <View className="h-28">
                         <FlatList
                             data={filterTemplates()}
                             renderItem={({item, index}) => (
                                 <Pressable
-                                    onPress={() => startTemplate(item)}
-                                    className="px-1"
+                                    onPress={() =>
+                                    {
+                                        startTemplate(item);
+                                    }}
+                                    className="px-1 border border-gray-300 rounded-lg"
                                 >
-                                    <Text className="pl-3 hover:outline hover:outline-1 rounded-md">
+                                    <Text className="pl-4 text-lg">
                                         {item.name}
                                     </Text>
                                 </Pressable>
@@ -74,7 +78,6 @@ export default function StartCurrentSession({Icon}: Props)
                                 <Text className="px-1 pl-3 ">No Templates</Text>
                             }
                             className="flex-1 py-1"
-                            initialNumToRender={3}
                         />
                     </View>
                 </View>
