+++ packages/native-frontend/components/templates/ImportAsTemplate.tsx
@@ -9,7 +9,7 @@ import {
     TouchableWithoutFeedback,
     TextInput,
 } from "react-native";
-import ModalTemplate from "../ModalTemplate";
+import ModalTemplate from "../UtilComponents/ModalTemplate";
 
 type Props = {
     fromSession: boolean;
@@ -66,7 +66,7 @@ export default function ImportAsTemplate({fromSession, Icon}: Props)
                     )}
                     <View className="flex-row justify-center items-center">
                         <Pressable onPress={() => setModalVisible(false)}>
-                            <Text className="text-2xl px-3 py-1 rounded-xl outline outline-1 outline-gray-300 bg-gray-200 hover:bg-gray-300">
+                            <Text className="text-2xl px-3 py-1 rounded-xl outline outline-1 outline-gray-300 bg-gray-200">
                                 Cancel
                             </Text>
                         </Pressable>
@@ -74,7 +74,7 @@ export default function ImportAsTemplate({fromSession, Icon}: Props)
                         <Pressable
                             onPress={() => saveTemplate(setModalVisible)}
                         >
-                            <Text className="text-2xl px-3 py-1 rounded-xl outline outline-1 outline-gray-300 bg-gray-200 hover:bg-gray-300">
+                            <Text className="text-2xl px-3 py-1 rounded-xl outline outline-1 outline-gray-300 bg-gray-200">
                                 Save
                             </Text>
                         </Pressable>
